Create Task2E.py
Not completed. At the moment, only outputting the graph for the first station in the list (need 5 graphs). Not letting me exit terminal, so the code is trying to continue to run. I understand why and need to fix this. The condition hasn't been added yet (for the 5 stations with greatest level), just want to output 5 graphs first.